# modules/timekeeping/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Timekeeper API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /punches:
    get:
      operationId: listPunches
      tags: [punches]
      summary: List punch events
      parameters:
        - in: query
          name: employeeId
          schema:
            type: ["integer", "null"]
        - in: query
          name: from
          schema:
            type: ["string", "null"]
            format: date
        - in: query
          name: to
          schema:
            type: ["string", "null"]
            format: date
        - in: query
          name: type
          schema:
            type: ["string", "null"]
            enum: [clock_in, clock_out, break_start, break_end]
      responses:
        "200":
          description: List of punch events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Punch"
    post:
      operationId: createPunch
      tags: [punches]
      summary: Record a new punch event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePunchBody"
      responses:
        "201":
          description: Punch recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Punch"
...
  /timesheets/{id}/leave-summary:
    get:
      operationId: getTimesheetLeaveSummary
      tags: [timesheets]
      summary: Get leave hours summary for a timesheet period
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leave summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimesheetLeaveSummary"

  # ─── KIOSK ────────────────────────────────────────────────────────────────
...
  /timesheets:
    get:
      operationId: listTimesheets
      tags: [timesheets]
      summary: List timesheets
      parameters:
        - in: query
          name: employeeId
          schema:
            type: ["integer", "null"]
        - in: query
          name: status
          schema:
            type: ["string", "null"]
            enum: [draft, submitted, approved, rejected]
        - in: query
          name: from
          schema:
            type: ["string", "null"]
            format: date
        - in: query
          name: to
          schema:
            type: ["string", "null"]
            format: date
      responses:
        "200":
          description: List of timesheets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Timesheet"
    post:
      operationId: createTimesheet
      tags: [timesheets]
      summary: Create a timesheet for a period
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimesheetBody"
      responses:
        "201":
          description: Created timesheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
  /timesheets/{id}:
    get:
      operationId: getTimesheet
      tags: [timesheets]
      summary: Get timesheet by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Timesheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
        "404":
          description: Not found
    patch:
      operationId: updateTimesheet
      tags: [timesheets]
      summary: Update timesheet
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTimesheetBody"
      responses:
        "200":
          description: Updated timesheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
  /timesheets/{id}/submit:
    post:
      operationId: submitTimesheet
      tags: [timesheets]
      summary: Submit a timesheet for approval
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submitted timesheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
  /timesheets/{id}/approve:
    post:
      operationId: approveTimesheet
      tags: [timesheets]
      summary: Approve a submitted timesheet
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved timesheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
  /timesheets/{id}/reject:
    post:
      operationId: rejectTimesheet
      tags: [timesheets]
      summary: Reject a submitted timesheet
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectTimesheetBody"
      responses:
        "200":
          description: Rejected timesheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
  /timesheets/{id}/recalculate:
    post:
      operationId: recalculateTimesheetHours
      tags: [timesheets]
      summary: Recalculate hours for a draft timesheet from current punch data
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recalculated timesheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
        "409":
          description: Timesheet is not in draft status
  /timesheets/{id}/punches:
    get:
      operationId: listTimesheetPunches
      tags: [timesheets]
      summary: List punches for a specific timesheet period (owner or admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Array of punches within the timesheet period
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Punch"
        "403":
          description: Not authorized to view this timesheet's punches
        "404":
          description: Timesheet not found

  # ─── CERTIFICATIONS ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Punch:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        type:
          type: string
          enum: [clock_in, clock_out, break_start, break_end]
        punchedAt:
          type: string
          format: date-time
        timezone:
          type: string
        note:
          type: ["string", "null"]
        source:
          type: string
          enum: [kiosk, web, admin, api]
        isEdited:
          type: boolean
        editNote:
          type: ["string", "null"]
        costCode:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - employeeId
        - type
        - punchedAt
        - timezone
        - source
        - isEdited
        - createdAt
        - updatedAt
    CreatePunchBody:
      type: object
      properties:
        employeeId:
          type: integer
        type:
          type: string
          enum: [clock_in, clock_out, break_start, break_end]
        punchedAt:
          type: ["string", "null"]
          format: date-time
        timezone:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        source:
          type: string
          enum: [kiosk, web, admin, api]
        costCode:
          type: ["string", "null"]
      required:
        - employeeId
        - type
        - source
...
    Timesheet:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        status:
          type: string
          enum: [draft, submitted, approved, rejected]
        totalHours:
          type: number
        regularHours:
          type: number
        overtimeHours:
          type: number
        rejectionNote:
          type: ["string", "null"]
        submittedAt:
          type: ["string", "null"]
          format: date-time
        reviewedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - employeeId
        - periodStart
        - periodEnd
        - status
        - totalHours
        - regularHours
        - overtimeHours
        - createdAt
        - updatedAt
    CreateTimesheetBody:
      type: object
      properties:
        employeeId:
          type: integer
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
      required:
        - employeeId
        - periodStart
        - periodEnd
    UpdateTimesheetBody:
      type: object
      properties:
        totalHours:
          type: number
        regularHours:
          type: number
        overtimeHours:
          type: number
    RejectTimesheetBody:
      type: object
      properties:
        rejectionNote:
          type: string
      required:
        - rejectionNote

    # ─── CERTIFICATION ──────────────────────────────────────────────────────
...
    LeaveEntry:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        date:
          type: string
          format: date
        leaveType:
          type: string
          enum: [pto, sick, holiday, bereavement, other]
        hours:
          type: number
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - employeeId
        - date
        - leaveType
        - hours
        - createdAt
        - updatedAt
...
    TimesheetLeaveSummary:
      type: object
      properties:
        workedHours:
          type: number
        leaveHours:
          type: number
        totalAccountedHours:
          type: number
        leaveEntries:
          type: array
          items:
            $ref: "#/components/schemas/LeaveEntry"
      required:
        - workedHours
        - leaveHours
        - totalAccountedHours
        - leaveEntries